contributors/adobe/submitted/shapes/basic-shapes/shape-outside-computed-shape-001.html

<!DOCTYPE html>
<html>
    <head>
        <title>Shape Outside Basic Shape Computed Viewport Relative Lengths</title>
        <link rel="author" title="Adobe" href="http://html.adobe.com/">
        <link rel="author" title="Bear Travis" href="mailto:betravis@adobe.com">
        <link rel="help" href="http://dev.w3.org/csswg/css-shapes/#shape-outside-property">
        <link rel="help" href="http://dev.w3.org/csswg/css-cascade/#computed">
        <meta name="assert" content="The basic shape can contain relative length formats, which resolve to the computed (absolute) length value">
        <meta name="flags" content="DOM">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/parsing-utils.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script type="text/javascript">
            // viewport relative units: vw, vh, vmin, vmax
            var units = ['vw', 'vh', 'vmin', 'vmax'];
            var div = document.createElement('div');
            document.body.appendChild(div);
            var resolveds = {};
            units.forEach(function(unit) {
                div.style.width = '1' + unit;
                var s = getComputedStyle(div);
                resolveds[unit] = parseFloat(s.width);
            });
            document.body.removeChild(div);

            function fillArray(string, length) {
                return Array.apply(null, new Array(length)).map(String.prototype.valueOf, string);
            }

            function testUnit(unit) {
                var relativeLength = '1' + unit;
                var usedLength = resolveds[unit] + 'px';

                var input = 'circle(' + fillArray('1' + unit, 3).join(', ') + ')';
                var output = 'circle(' + fillArray(resolveds[unit] + 'px', 3).join(', ') + ')';

                test_computed_style(input, output);
            }

            var tests = units.map(function(unit) {
                return [unit + ' units', unit];
            });

            generate_tests(testUnit, tests);
        </script>
    </body>
</html>
